refactor: lots of name changes and minor fixes
